Changes made to index.html and styles.css
I added the table along with styling to give more look. I centered the table via making the table's width 100% bigger. I put the table in a flex-box so it could have a unique color as opposed to the background color. I also made subtle changes to the font-families of certain text.