Finalization and completion
I added more info to the table data. I made sure to add a different font-family to the data under "Where". I then ran my code through a HTML and CSS validator to make sure that there were no errors. 

This is the finalized product, and a example of a table being used in HTML with a bit of CSS being used for visual changes to the webpage and the table itself.