Final Additional Changes
I needed to make a few more changes to the webpage and the table itself. I added the days of the week and I added a table header that takes up all columns. I added more data to the table as well. I made a couple more font-family changes as well to help with visuals. I of course then re-ran it through a HTML and CSS validator to make sure my changes didn't negatively impact the code.

Everything is fine now with the table and this is the actual final look.